Fix types.

+++ probabilistic_algorithms/reservoir_sampling.h
@@ -13,7 +13,7 @@ template<typename T>
 std::vector<T> reservoir_sampling(const std::vector<T> &input_stream, size_t reservoir_size)
 {
 	std::vector<T> reservoir(reservoir_size);
-	int element_index{};
+	size_t element_index{};
 	std::random_device random_device;
 	std::mt19937 generator(random_device());
 	for (const auto &element: input_stream) {
@@ -21,7 +21,7 @@ std::vector<T> reservoir_sampling(const std::vector<T> &input_stream, size_t res
 			reservoir[element_index] = element;
 		else {
 			std::uniform_int_distribution<> distribution(0, element_index);
-			int reservoir_index = distribution(generator);
+			size_t reservoir_index = distribution(generator);
 			if (reservoir_index < reservoir_size)
 				reservoir[reservoir_index] = element;
 		}
@@ -41,7 +41,7 @@ template<typename T>
 std::vector<T> reservoir_sampling(const Node<T> *head, size_t reservoir_size)
 {
 	std::vector<Node<T> *> reservoir(reservoir_size);
-	int element_index{};
+	size_t element_index{};
 	std::random_device random_device;
 	std::mt19937 generator(random_device());
 	for (auto current_node = head; current_node; current_node = current_node->next, ++element_index) {
@@ -49,7 +49,7 @@ std::vector<T> reservoir_sampling(const Node<T> *head, size_t reservoir_size)
 			reservoir[element_index] = current_node->value;
 		else {
 			std::uniform_int_distribution<> distribution(0, element_index);
-			int reservoir_index = distribution(generator);
+			size_t reservoir_index = distribution(generator);
 			if (reservoir_index < reservoir_size)
 				reservoir[reservoir_index] = current_node->value;
 		}
